refactor(consultant): split ConsultantFarmPage, fix atomic send + review bugs
Split the 1917-line ConsultantFarmPage into 16 focused colocated files
(parent orchestration + 14 components + 2 config/helper modules), and fix
four review findings on the farm/plan flow:

- P1: first-plan send now routes through FertilizerPlanService.sendPlan
  (send_fertilizer_plan RPC) when a Petiole Review is pending, so the
  plan<->review link is established and the review flips to 'reviewed'
  in one atomic transaction. The prior createPlan + updateTriageReview
  sequence was non-atomic and left petiole_triage_id null, breaking
  get_farmer_recommendations and leaving the queue stale on partial
  failure. Ad-hoc plans (no pending review) still use createPlan.
- P2: reject quantity <= 0 before send so the first-plan path can't
  persist an item the edit path's RPC validator would later reject.
- P2: add a staleness guard to the farmer profile load so navigating
  between [farmerId] segments can't overwrite farmer B's state with
  farmer A's in-flight response.
- P2: VisitHistory time label now derives from the record's timestamptz
  createdAt instead of parsing the date-only visit_date as UTC midnight
  (which fabricated an identical 5:30am for every visit in IST).
- document-service: drop unused `data` destructure and empty `// ...`
  else-if branch in ensureBucket.
