+++ src/app/consultant/farmers/[farmerId]/farms/[farmId]/components/SectionLabel.tsx
@@ -1,0 +1,7 @@
+export function SectionLabel({ children }: { children: React.ReactNode }) {
+  return (
+    <h2 className="text-xs font-semibold uppercase tracking-wider text-muted-foreground">
+      {children}
+    </h2>
+  )
+}
